+++ callisto_core/delivery/forms.py
@@ -43,7 +43,7 @@ class ReportBaseForm(
 
     def save(self, *args, **kwargs):
         if self.data.get("key"):
-            self.view.storage.set_secret_key(self.data['key'])
+            self.view.storage.set_passphrase(self.data['key'])
         return super().save(*args, **kwargs)
 
     class Meta: