+++ callisto_core/tests/delivery/test_views.py
@@ -18,11 +18,11 @@ class LegacyStorageFormatTest(test_base.ReportFlowHelper):
         super().setUp()
         self.report = models.Report.objects.create()
         legacy_storage = {'data': {'catte': 'good'}}
-        self.report.encrypt_report(legacy_storage, self.secret_key)
-        self.client_set_secret_key()
+        self.report.encrypt_report(legacy_storage, self.passphrase)
+        self.client_set_passphrase()
         self.client_post_answer_question()  # prompt code to initialize storage
         self.report.refresh_from_db()
-        self.storage = self.report.decrypted_report(self.secret_key)
+        self.storage = self.report.decrypted_report(self.passphrase)
 
     def test_form_data_populated(self):
         self.assertTrue(self.storage.get('wizard_form_serialized', False))
@@ -56,13 +56,13 @@ class NewReportFlowTest(test_base.ReportFlowHelper):
 
     def test_report_creation_adds_key_to_session(self):
         self.assertEqual(
-            self.client.session.get('secret_key'),
+            self.client.session.get('passphrase'),
             None,
         )
         self.client_post_report_creation()
         self.assertEqual(
-            self.client.session.get('secret_key'),
-            self.secret_key,
+            self.client.session.get('passphrase'),
+            self.passphrase,
         )
 
     def test_access_form_rendered_when_no_key_in_session(self):
@@ -71,20 +71,20 @@ class NewReportFlowTest(test_base.ReportFlowHelper):
         page_1_path = reverse(
             'report_update', kwargs={
                 'step': 0, 'uuid': uuid})
-        self.client_clear_secret_key()
+        self.client_clear_passphrase()
 
         response = self.client.get(page_1_path)
         form = response.context['form']
 
         self.assertIsInstance(form, forms.ReportAccessForm)
 
-    def test_can_reenter_secret_key(self):
+    def test_can_reenter_passphrase(self):
         response = self.client_post_report_creation()
         uuid = response.context['report'].uuid
         page_1_path = reverse(
             'report_update', kwargs={
                 'step': 0, 'uuid': uuid})
-        self.client_clear_secret_key()
+        self.client_clear_passphrase()
 
         response = self.client_post_report_access(page_1_path)
         self.assertRedirects(response, page_1_path)
@@ -92,7 +92,7 @@ class NewReportFlowTest(test_base.ReportFlowHelper):
     def test_access_form_returns_correct_report(self):
         response = self.client_post_report_creation()
         uuid = response.context['report'].uuid
-        self.client_clear_secret_key()
+        self.client_clear_passphrase()
 
         response = self.client_post_report_access(
             response.redirect_chain[0][0])
@@ -101,9 +101,9 @@ class NewReportFlowTest(test_base.ReportFlowHelper):
 
     def test_report_not_accessible_with_incorrect_key(self):
         response = self.client_post_report_creation()
-        self.client_clear_secret_key()
+        self.client_clear_passphrase()
 
-        self.secret_key = 'wrong key'
+        self.passphrase = 'wrong key'
         response = self.client_post_report_access(
             response.redirect_chain[0][0])
         form = response.context['form']
@@ -117,15 +117,15 @@ class ReportMetaFlowTest(test_base.ReportFlowHelper):
     def test_report_action_no_key(self):
         self.client_post_report_creation()
         self.assertTrue(self.report.pk)
-        self.client_clear_secret_key()
+        self.client_clear_passphrase()
         self.client_post_report_delete()
         self.assertTrue(self.assert_report_exists())
 
     def test_report_action_invalid_key(self):
         self.client_post_report_creation()
         self.assertTrue(self.report.pk)
-        self.client_clear_secret_key()
-        self.secret_key = 'wrong key'
+        self.client_clear_passphrase()
+        self.passphrase = 'wrong key'
         self.client_post_report_delete()
         self.assertTrue(self.assert_report_exists())