+++ callisto_core/utils/settings.py
@@ -1,7 +1,7 @@
 import os
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-SECRET_KEY = os.getenv("SECRET_KEY", default='secret key')
+passphrase = os.getenv("passphrase", default='secret key')
 DEBUG = True
 
 ROOT_URLCONF = "callisto_core.utils.urls"