+++ docs/HISTORY.md
@@ -6,9 +6,9 @@
 * scope stored passphrase to current report uuid
 * rename various view methods
     * `KeyResetTemplatePartial` => `PassphraseClearingPartial`
-    * `clear_secret_key` => `clear_passphrases`
-    * `set_secret_key` => `set_passphrase`
-    * `secret_key` => `passphrase`
+    * `clear_passphrase` => `clear_passphrases`
+    * `set_passphrase` => `set_passphrase`
+    * `passphrase` => `passphrase`
 * remove `report_and_key_present`, as `passphrase` needs a report now
 
     https://github.com/project-callisto/callisto-core/pull/313